centralize use of transaction cache to LocalCacheTransactionRedis.
